Suppress "insecure origin" warning if script is loaded from a Firefox WebExtension
In the same vein as "chrome-extension:", the "moz-extension:" protocol is used by Mozilla Firefox to access files within locally installed browser extensions. As such it is not susceptible to MitM attacks.

Point of note: there seems to be a W3C proposal being worked on that will unify the browser protocol for files in local extensions as "browserext:". However, since that is not in use anywhere yet and the proposal is not official at this point, I am not including it in this pull request.